[FIX] Add support for >, < in version specifiers
Do not use `shell=True` when running `pip` commands.

Prior to this commit, `shell=True` was used when running `pip` commands.
However, `shell=True` does not quote arguments properly.  This prevents,
for example, installing packages with version specifiers containing `<`
or `>`, e.g., `foo>=0.1`.

As of this commit, `shell=False` is used when running `pip` commands.
This adds support for installing packages with version specifiers
containing `<` or `>`.